Added notifications to Admin page.
Updated how frontend errors are arranged and corresponding tests to be
able to specify which mesages should be displayed as notification. For
example the fact that all articles are succesfully loaded should not be
displayed as a notification but the fact that an article was successfully created should be.

Added a selenium test which creates and article and asserts that the
notification is displayed notifying the user.

Updated the caching to only work in ci and prod.